# openapi.json
{
  "components" : {
    "schemas" : {
...
      "BusinessView" : {
        "properties" : {
          "country" : {
            "type" : [ "string", "null" ]
          },
          "creditDefaultDueDays" : {
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "creditLimitEnforced" : {
            "type" : "boolean"
          },
          "creditOverdueDays" : {
            "format" : "int32",
            "type" : "integer"
          },
          "creditRequiresCustomer" : {
            "type" : "boolean"
          },
          "currency" : {
            "type" : [ "string", "null" ]
          },
          "dayCutoff" : {
            "type" : [ "string", "null" ]
          },
          "dayRuleEffectiveFrom" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "dayRules" : {
            "items" : {
              "$ref" : "#/components/schemas/DayRuleView"
            },
            "type" : [ "array", "null" ]
          },
          "defaultLocale" : {
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "inventoryMode" : {
            "type" : [ "string", "null" ]
          },
          "modules" : {
            "additionalProperties" : {
              "type" : "boolean"
            },
            "type" : [ "object", "null" ]
          },
          "name" : {
            "type" : [ "string", "null" ]
          },
          "posViews" : {
            "items" : {
              "type" : "string"
            },
            "type" : [ "array", "null" ]
          },
          "shiftNoteThresholdMinor" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "shiftRequired" : {
            "type" : "boolean"
          },
          "status" : {
            "type" : [ "string", "null" ]
          },
          "timezone" : {
            "type" : [ "string", "null" ]
          },
          "type" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "creditLimitEnforced", "creditOverdueDays", "creditRequiresCustomer", "id", "shiftRequired" ],
        "type" : "object"
      },
...
      "CategoryInput" : {
        "properties" : {
          "active" : {
            "type" : [ "boolean", "null" ]
          },
          "name" : {
            "type" : [ "string", "null" ]
          }
        },
        "type" : "object"
      },
      "CategoryView" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : [ "string", "null" ]
          },
          "rev" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "active", "id", "rev" ],
        "type" : "object"
      },
...
      "DayRuleView" : {
        "properties" : {
          "dayCutoff" : {
            "type" : [ "string", "null" ]
          },
          "from" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "timezone" : {
            "type" : [ "string", "null" ]
          }
        },
        "type" : "object"
      },
...
      "ExpenseCategoryInput" : {
        "properties" : {
          "active" : {
            "type" : [ "boolean", "null" ]
          },
          "name" : {
            "type" : [ "string", "null" ]
          }
        },
        "type" : "object"
      },
      "ExpenseCategoryView" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "key" : {
            "type" : [ "string", "null" ]
          },
          "name" : {
            "type" : [ "string", "null" ]
          },
          "rev" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "active", "id", "rev" ],
        "type" : "object"
      },
...
      "ImportBody" : {
        "properties" : {
          "rows" : {
            "items" : {
              "$ref" : "#/components/schemas/ImportRow"
            },
            "type" : [ "array", "null" ]
          }
        },
        "type" : "object"
      },
      "ImportResult" : {
        "properties" : {
          "dryRun" : {
            "type" : "boolean"
          },
          "rows" : {
            "items" : {
              "$ref" : "#/components/schemas/ImportRowResult"
            },
            "type" : [ "array", "null" ]
          },
          "summary" : {
            "$ref" : "#/components/schemas/ImportSummary"
          }
        },
        "required" : [ "dryRun" ],
        "type" : "object"
      },
...
      "PageResponseProductView" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/ProductView"
            },
            "type" : [ "array", "null" ]
          },
          "last" : {
            "type" : "boolean"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "total" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "last", "page", "size", "total" ],
        "type" : "object"
      },
...
      "ProductInput" : {
        "properties" : {
          "active" : {
            "type" : [ "boolean", "null" ]
          },
          "barcode" : {
            "type" : [ "string", "null" ]
          },
          "categoryId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "color" : {
            "type" : [ "string", "null" ]
          },
          "costMinor" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "isQuick" : {
            "type" : [ "boolean", "null" ]
          },
          "minStockMilli" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "name" : {
            "type" : [ "string", "null" ]
          },
          "priceMinor" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "pricing" : {
            "type" : [ "string", "null" ]
          },
          "quickPosition" : {
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "shortCode" : {
            "type" : [ "string", "null" ]
          },
          "trackStock" : {
            "type" : [ "boolean", "null" ]
          },
          "unit" : {
            "type" : [ "string", "null" ]
          },
          "variant" : {
            "type" : [ "string", "null" ]
          }
        },
        "type" : "object"
      },
      "ProductView" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "barcode" : {
            "type" : [ "string", "null" ]
          },
          "categoryId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "color" : {
            "type" : [ "string", "null" ]
          },
          "costMinor" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "isQuick" : {
            "type" : "boolean"
          },
          "minStockMilli" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "name" : {
            "type" : [ "string", "null" ]
          },
          "priceMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pricing" : {
            "type" : [ "string", "null" ]
          },
          "quickPosition" : {
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "rev" : {
            "format" : "int64",
            "type" : "integer"
          },
          "shortCode" : {
            "type" : [ "string", "null" ]
          },
          "stockMilli" : {
            "format" : "int64",
            "type" : "integer"
          },
          "trackStock" : {
            "type" : "boolean"
          },
          "unit" : {
            "type" : [ "string", "null" ]
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "variant" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "active", "id", "isQuick", "priceMinor", "rev", "stockMilli", "trackStock" ],
        "type" : "object"
      },
...
      "UpdateBusiness" : {
        "properties" : {
          "clearCreditDefaultDueDays" : {
            "type" : [ "boolean", "null" ]
          },
          "clearShiftNoteThreshold" : {
            "type" : [ "boolean", "null" ]
          },
          "creditDefaultDueDays" : {
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "creditLimitEnforced" : {
            "type" : [ "boolean", "null" ]
          },
          "creditOverdueDays" : {
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "creditRequiresCustomer" : {
            "type" : [ "boolean", "null" ]
          },
          "currency" : {
            "type" : [ "string", "null" ]
          },
          "dayCutoff" : {
            "type" : [ "string", "null" ]
          },
          "defaultLocale" : {
            "type" : [ "string", "null" ]
          },
          "modules" : {
            "additionalProperties" : {
              "type" : "boolean"
            },
            "type" : [ "object", "null" ]
          },
          "name" : {
            "type" : [ "string", "null" ]
          },
          "posViews" : {
            "items" : {
              "type" : "string"
            },
            "type" : [ "array", "null" ]
          },
          "shiftNoteThresholdMinor" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "shiftRequired" : {
            "type" : [ "boolean", "null" ]
          },
          "timezone" : {
            "type" : [ "string", "null" ]
          },
          "type" : {
            "type" : [ "string", "null" ]
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/b/{businessId}" : {
      "delete" : {
        "operationId" : "delete_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "202" : {
            "description" : "Accepted"
          }
        },
        "tags" : [ "business-controller" ]
      },
      "get" : {
        "operationId" : "get",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BusinessView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "business-controller" ]
      },
      "put" : {
        "operationId" : "update_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateBusiness"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BusinessView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "business-controller" ]
      }
    },
...
    "/api/b/{businessId}/categories" : {
      "get" : {
        "operationId" : "categories_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CategoryView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "catalog-controller" ]
      }
    },
    "/api/b/{businessId}/categories/{categoryId}" : {
      "put" : {
        "operationId" : "upsertCategory_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "categoryId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CategoryInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "catalog-controller" ]
      }
    },
...
    "/api/b/{businessId}/expense-categories" : {
      "get" : {
        "operationId" : "categories",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ExpenseCategoryView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "cash-controller" ]
      }
    },
    "/api/b/{businessId}/expense-categories/{categoryId}" : {
      "put" : {
        "operationId" : "upsertCategory",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "categoryId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExpenseCategoryInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpenseCategoryView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "cash-controller" ]
      }
    },
...
    "/api/b/{businessId}/products" : {
      "get" : {
        "operationId" : "search",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "quick",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "includeInactive",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "updatedSince",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseProductView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "catalog-controller" ]
      }
    },
    "/api/b/{businessId}/products/barcode/{code}" : {
      "get" : {
        "operationId" : "byBarcode",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "catalog-controller" ]
      }
    },
    "/api/b/{businessId}/products/import" : {
      "post" : {
        "operationId" : "importProducts",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "dryRun",
          "required" : false,
          "schema" : {
            "default" : true,
            "type" : "boolean"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ImportBody"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImportResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "catalog-controller" ]
      }
    },
    "/api/b/{businessId}/products/{productId}" : {
      "delete" : {
        "operationId" : "deactivate",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "productId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "catalog-controller" ]
      },
      "get" : {
        "operationId" : "get_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "productId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "catalog-controller" ]
      },
      "put" : {
        "operationId" : "upsert_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "productId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProductInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "catalog-controller" ]
      }
    },
    "/api/b/{businessId}/products/{productId}/history" : {
      "get" : {
        "operationId" : "history",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "productId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ProductHistoryEntry"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "catalog-controller" ]
      }
    },
...
    "/api/b/{businessId}/reports/products" : {
      "get" : {
        "operationId" : "products",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : "revenue",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Product"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "report-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}